# Use SSE2 intrinsics to find longest common prefix