This PR adds SSE2 intrinsics to the method to find the longest common prefix. Benchmark results:

Before
```
Testing !/redwood/correctness/unit/RedwoodRecordRef
DeltaTest() on random large records 0.163069 M/s  3538.33 MB/s
DeltaTest() on random small records 163.069 M/s  6376.88 MB/s
5700000000 getCommonPrefixLen(skip=50) 515.447 M/s
5700000000 getCommonPrefixLen(skip=0) 114.42 M/s
1300000000 writeDelta(commonPrefix=57) 101.44 M/s
130000000 writeDelta() 44.4518 M/s
```

After
```
Testing !/redwood/correctness/unit/RedwoodRecordRef
DeltaTest() on random large records 0.317222 M/s  7141.03 MB/s
DeltaTest() on random small records 317.222 M/s  12397 MB/s
5700000000 getCommonPrefixLen(skip=50) 441.804 M/s
5700000000 getCommonPrefixLen(skip=0) 118.944 M/s
1300000000 writeDelta(commonPrefix=57) 101.98 M/s
130000000 writeDelta() 53.5847 M/s
```

Run on an r5d.2xl.