@Daniel-B-Smith 2020-06-29
Since SSE2 instructions are mandatory for x86_64, this is portable on those platforms. Obviously, this won't work on 32 bit platforms or anything not Intel. I'm happy to add whatever appropriate ifdefs.