@Daniel-B-Smith 2020-06-29
Whoops. Fixed the 16/128 confusion. I guess I only ran the performance test and not the correctness test.