@sfc-gh-kmakino 2020-06-29
> > Since SSE2 instructions are mandatory for x86_64, this is portable on those platforms. Obviously, this won't work on 32 bit platforms or anything not Intel. I'm happy to add whatever appropriate ifdefs.
> 
> How about ARM aarm64? Will it break ARM compilation?
> FYI, @kaomakino

The intrinsics header file isn't available on aarch64, which needs to be handled properly.
All of SSE instructions are mapped to NEON instructions by sse2neon, so this PR shouldn't break aarch64.
However, ideally we should wrap those processor specific SIMD instructions in Platform.{cpp,h} one day (like what we do for CRC or TSC today).