@Daniel-B-Smith 2020-06-30
After turning on AVX instructions, the existing code ends up being faster than code using `_mm_cmpeq_epi8` which is in turn faster than the code using `_mm_cmpestri`. My guess is the misaligned loads are what is causing the `_mm_cmpeq_epi8` version to be slower. If we could ensure that both strings had the same alignment, it would be easier to handle the misaligned bytes first and only operate on aligned data. Unfortunately, I suspect the strings will have different alignments much of the time. I'm going to close this as there is no obvious path forward.

For reference, `_mm_loadu_si128` was faster than using `_mm_lddqu_si128`. Our best guess is that is the case because most of the time we're not reading across a cache line boundary.